Move shift request handler into blueprint
