+++ tests/test_absence_requests_blueprint.py
@@ -28,3 +28,10 @@ def test_leave_handler_is_no_longer_implemented_in_legacy_module():
 
     source = (Path(__file__).parents[1] / "app.py").read_text()
     assert "def leave():" not in source
+
+
+def test_shift_request_handler_is_no_longer_implemented_in_legacy_module():
+    from pathlib import Path
+
+    source = (Path(__file__).parents[1] / "app.py").read_text()
+    assert "def requests_page():" not in source
